Switch from matplotlib to NEURON plots
A fresh install of NEURON on an OS 10.7 machine did not have matplotlib
available.  To reduce dependencies and simplify the install process,
the matplotlib code was removed and NEURON plotting code added.  Note
that you will need to run the GUI version of NEURON to see the plots
(see README.mkd).