Remove extraneous columns from sweep csv headers
The sweep csv header contained all of the variables that depended on
sweep_param and threshold_param, but the remaining rows only contained
fields with numeric values.  This was because some of the variables also
depended on position.  To ensure the headers match the numeric data, the
columns that depend on position were removed from the list of headers.